+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Date format: (YYYY-MM-DD)
 * Bumped the Bukkit dependency to Spigot 1.18.1. This version is not compatible with older Bukkit versions.
 * Bumped the ProtocolLib dependency to version 4.8.0.
 * Similar to how we update the nearby signs when a player joins the server, we now also update the nearby signs for players that are already online when the plugin is enabled.
+* We now load all plugin classes right away. This avoids class loading issues when the plugin jar is dynamically replaced while the plugin is still running.
 * Internal: Updated to Gradle 7.3.3.
 
 ## v2.8.0 Release (2021-08-08)